S6/poetry test/pyproject.toml
[tool.poetry]
name = "poetry test"
version = "0.1.0"
description = "Test 1"
authors = ["RichardJacomo <richard.jacomo23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ipdb = "^0.13.13"
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
